move otelinit to utils
