+++ app/utils/otelInit.go
@@ -1,13 +1,12 @@
-package main
+package utils
 
 import (
 	"context"
-	"os"
+	"log/slog"
 	"time"
 
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
 	"go.opentelemetry.io/otel/sdk/resource"
-	"go.opentelemetry.io/otel/sdk/trace"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 )
@@ -16,11 +15,11 @@ import (
 // 1. define to where the data will be sent (for this case stdout), exporter are defined per signal
 // 2. Create the providers of the signals to use
 
-func initOpenTelemetry() *sdktrace.TracerProvider {
+func InitOpenTelemetry(otelUrl string, svcName string, logger *slog.Logger) *sdktrace.TracerProvider {
 	// define exporter for traces
 	exporter, err := otlptracehttp.New(
 		context.Background(),
-		otlptracehttp.WithEndpoint(os.Getenv("OTEL_COLLECTOR")+":4318"),
+		otlptracehttp.WithEndpoint(otelUrl+":4318"),
 		otlptracehttp.WithInsecure(),
 	)
 
@@ -34,14 +33,14 @@ func initOpenTelemetry() *sdktrace.TracerProvider {
 	tp := sdktrace.NewTracerProvider(
 		sdktrace.WithBatcher(
 			exporter,
-			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
-			trace.WithBatchTimeout(trace.DefaultScheduleDelay*time.Millisecond),
-			trace.WithMaxExportBatchSize(trace.DefaultMaxExportBatchSize),
+			sdktrace.WithMaxExportBatchSize(sdktrace.DefaultMaxExportBatchSize),
+			sdktrace.WithBatchTimeout(sdktrace.DefaultScheduleDelay*time.Millisecond),
+			sdktrace.WithMaxExportBatchSize(sdktrace.DefaultMaxExportBatchSize),
 		),
-		trace.WithResource(
+		sdktrace.WithResource(
 			resource.NewWithAttributes(
 				semconv.SchemaURL,
-				semconv.ServiceNameKey.String("user-service"),
+				semconv.ServiceNameKey.String(svcName),
 			),
 		),
 	)
